# Optimize StrongReferenceMessenger