This PR brings some of the optimizations and improvements from #44 to `StrongReferenceMessenger` too.
Specifically, this includes:

- Using double-dispatch through `MessageHandlerDispatcher` to avoid aliasing delegate types during broadcast.
- Type specialization for `Unit` (used when sending messages through the default channel)
- Specialization for `IRecipient<TMessage>` to avoid the double delegate dispatch
- Other general tweaks and code refactoring

## Benchmarks

| Method |     Mean |     Error |    StdDev | Ratio | Allocated |
|------- |---------:|----------:|----------:|------:|----------:|
|   main | 5.586 ms | 0.0317 ms | 0.0264 ms |  1.00 |         - |
|     **PR** | **4.013 ms** | 0.0074 ms | 0.0058 ms |  **0.71** |         - |

Overall this results in a ~30% speedup 🚀